DickerBotHost/requirements.txt
pyinstaller
pyserial
PyQt5
websockets